[bug 875313] Use compile-linted-mo.sh in update script
This changes the update_locales part of the update script which we
use for deployments to use compile-linted-mo.sh instead of compile-mo.sh.

This will cause the deploy logs to contain a list of the files that
were compiled and the ones that are BUSTED. For now, we'll have to check
that for .po files that have errors that need to be fixed.